Move interpolation module to own folder
