+++ setup.cfg
@@ -21,7 +21,7 @@ classifiers =
     Topic :: Scientific/Engineering
     Topic :: Software Development :: Build Tools
 project_urls =
-    Github = https://github.com/segsell/dcegm
+    Github = https://github.com/OpenSourceEconomics/dcegm
 
 [options]
 packages = find:
