+++ src/dcegm/interpolation/interp1d.py
